convert run.sh to run
Importantly, it puts all the possible commands into a single hyperfine
indication, which can help you get a relative comparison in the final
report. Also, use a normal number of runs and warmups to get more
statistically significant results.

The use of Zsh is for portability with associative arrays (macOS's
/bin/bash is still too old on most systems).